Extract Emoji class from EmojiAdapter
